{"schema":"decodex/commit/1","summary":"Modularize post-review and intake render boundaries","authority":"manual"}
