+++ apps/decodex/src/orchestrator/status/post_review/lanes.rs
@@ -1,263 +1,8 @@
-use crate::{
-	orchestrator::status::{
-		post_review,
-		post_review::{
-			HashMap, OperatorLoopStatus, OperatorPostReviewLaneStatus, OperatorStatusSnapshot,
-			PostReviewLaneBuildContext, PostReviewLaneClassification, PostReviewLaneDecision,
-			PostReviewLaneSnapshot, PullRequestReviewStateInspector, ServiceConfig, StateStore,
-			TrackerIssue, WorkflowDocument, WorktreeMapping,
-		},
-	},
-	prelude::Result,
-	tracker,
+mod build;
+mod metadata;
+mod status;
+
+pub(crate) use self::{
+	build::build_post_review_lane_statuses_from_worktree_issues,
+	metadata::hydrate_worktree_issue_metadata,
 };
-
-pub(crate) fn build_post_review_lane_statuses_from_worktree_issues<I>(
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	review_state_inspector: &I,
-	worktree_issues: Vec<(WorktreeMapping, TrackerIssue)>,
-) -> Result<Vec<OperatorPostReviewLaneStatus>>
-where
-	I: PullRequestReviewStateInspector,
-{
-	let tracker_policy = workflow.frontmatter().tracker();
-	let success_state = tracker_policy.success_state();
-	let completed_state = tracker_policy.resolved_completed_state();
-	let lane_context = PostReviewLaneBuildContext {
-		project,
-		workflow,
-		state_store,
-		review_state_inspector,
-		success_state,
-		completed_state,
-	};
-	let mut lanes = Vec::new();
-
-	for (worktree, issue) in worktree_issues {
-		let Some(lane) = build_post_review_lane_status(&lane_context, issue, worktree)? else {
-			continue;
-		};
-
-		lanes.push(lane);
-	}
-
-	lanes.sort_by(|left, right| left.issue_identifier.cmp(&right.issue_identifier));
-
-	Ok(lanes)
-}
-
-pub(crate) fn hydrate_worktree_issue_metadata(
-	snapshot: &mut OperatorStatusSnapshot,
-	worktree_issues: &[(WorktreeMapping, TrackerIssue)],
-) {
-	let issues_by_id = worktree_issues
-		.iter()
-		.map(|(_, issue)| (issue.id.as_str(), issue))
-		.collect::<HashMap<_, _>>();
-
-	for worktree in &mut snapshot.worktrees {
-		let Some(issue) = issues_by_id.get(worktree.issue_id.as_str()) else {
-			continue;
-		};
-
-		worktree.issue_identifier = Some(issue.identifier.clone());
-		worktree.issue_state = Some(issue.state.name.clone());
-	}
-}
-
-pub(crate) fn build_post_review_lane_status<I>(
-	context: &PostReviewLaneBuildContext<'_, I>,
-	issue: TrackerIssue,
-	worktree: WorktreeMapping,
-) -> Result<Option<OperatorPostReviewLaneStatus>>
-where
-	I: PullRequestReviewStateInspector,
-{
-	if issue.state.name != context.success_state && issue.state.name != context.completed_state {
-		return Ok(None);
-	}
-
-	if let Some(reason) = post_review_lane_static_block_reason(&issue, context.workflow)? {
-		return Ok(Some(post_review::blocked_post_review_lane_status(
-			context.project,
-			&issue,
-			&worktree,
-			reason,
-		)));
-	}
-
-	let retry_budget_exhausted = post_review::issue_retry_budget_exhausted_for_worktree(
-		context.workflow,
-		context.state_store,
-		&issue.id,
-		worktree.worktree_path(),
-	)?;
-	let review_handoff = context.state_store.review_handoff_marker(
-		context.project.service_id(),
-		&issue.id,
-		worktree.branch_name(),
-	)?;
-
-	if issue.state.name == context.completed_state && review_handoff.is_none() {
-		return Ok(None);
-	}
-
-	let local_branch_name =
-		match post_review::worktree_checkout_branch_name(worktree.worktree_path()) {
-			Ok(local_branch_name) => local_branch_name,
-			Err(_error) => {
-				return Ok(Some(post_review::blocked_post_review_lane_status(
-					context.project,
-					&issue,
-					&worktree,
-					"worktree_checkout_branch_read_failed",
-				)));
-			},
-		};
-	let local_head_oid = match post_review::worktree_head_oid(worktree.worktree_path()) {
-		Ok(local_head_oid) => local_head_oid,
-		Err(_error) => {
-			return Ok(Some(post_review::blocked_post_review_lane_status(
-				context.project,
-				&issue,
-				&worktree,
-				"worktree_head_read_failed",
-			)));
-		},
-	};
-	let snapshot = PostReviewLaneSnapshot {
-		issue,
-		worktree,
-		review_handoff,
-		local_branch_name,
-		local_head_oid,
-	};
-	let mut classification = post_review::classify_post_review_lane_with_project(
-		&snapshot,
-		context.project,
-		context.workflow,
-		context.state_store,
-		context.review_state_inspector,
-	)?;
-
-	if retry_budget_exhausted {
-		classification = post_review::retry_budget_exhausted_post_review_lane_classification(
-			&snapshot,
-			context.project,
-			context.workflow,
-			context.review_state_inspector,
-			classification,
-		);
-	}
-
-	apply_active_ownership_warning_to_post_review_lane(
-		context.project,
-		context.success_state,
-		&snapshot,
-		&mut classification,
-	);
-
-	Ok(Some(post_review_lane_status_from_classification(
-		context.project,
-		context.state_store,
-		&snapshot,
-		classification,
-	)?))
-}
-
-pub(crate) fn apply_active_ownership_warning_to_post_review_lane(
-	project: &ServiceConfig,
-	success_state: &str,
-	snapshot: &PostReviewLaneSnapshot,
-	classification: &mut PostReviewLaneClassification,
-) {
-	if snapshot.review_handoff.is_none()
-		|| snapshot.issue.state.name != success_state
-		|| !snapshot.issue.labels_complete
-		|| snapshot.issue.has_label(&tracker::automation_active_label(project.service_id()))
-	{
-		return;
-	}
-	if classification.readback_warning.is_none() {
-		classification.readback_warning = Some(String::from("active_ownership_label_missing"));
-	}
-}
-
-pub(crate) fn post_review_lane_status_from_classification(
-	project: &ServiceConfig,
-	state_store: &StateStore,
-	snapshot: &PostReviewLaneSnapshot,
-	classification: PostReviewLaneClassification,
-) -> Result<OperatorPostReviewLaneStatus> {
-	let loop_status =
-		operator_post_review_loop_status(project, state_store, snapshot, classification.decision)?;
-
-	Ok(OperatorPostReviewLaneStatus {
-		project_id: project.service_id().to_owned(),
-		issue_id: snapshot.issue.id.clone(),
-		issue_identifier: snapshot.issue.identifier.clone(),
-		issue_state: snapshot.issue.state.name.clone(),
-		branch_name: snapshot.worktree.branch_name().to_owned(),
-		worktree_path: post_review::relative_worktree_path_for_path(
-			project,
-			snapshot.worktree.worktree_path(),
-		),
-		classification: classification.decision.as_str().to_owned(),
-		reason: classification.reason,
-		pr_url: classification.pr_url,
-		pr_head_sha: classification.pr_head_sha,
-		pr_state: classification.pr_state,
-		review_decision: classification.review_decision,
-		mergeable: classification.mergeable,
-		check_state: classification.check_state,
-		unresolved_review_threads: classification.unresolved_review_threads,
-		shadowed_by_current_lane: false,
-		readback_warning: classification.readback_warning,
-		readback_root_cause: classification.readback_root_cause,
-		loop_status,
-	})
-}
-
-pub(crate) fn operator_post_review_loop_status(
-	project: &ServiceConfig,
-	state_store: &StateStore,
-	snapshot: &PostReviewLaneSnapshot,
-	decision: PostReviewLaneDecision,
-) -> Result<Option<OperatorLoopStatus>> {
-	let Some(review_handoff) = snapshot.review_handoff.as_ref() else {
-		return Ok(None);
-	};
-	let default_review_phase = match decision {
-		PostReviewLaneDecision::ReadyToLand | PostReviewLaneDecision::WaitForReview => None,
-		_ => Some("repair"),
-	};
-
-	post_review::operator_loop_status_for_run(
-		project,
-		state_store,
-		&snapshot.issue.id,
-		review_handoff.run_id(),
-		review_handoff.attempt_number(),
-		default_review_phase,
-		None,
-	)
-	.map(Some)
-}
-
-pub(crate) fn post_review_lane_static_block_reason(
-	issue: &TrackerIssue,
-	workflow: &WorkflowDocument,
-) -> Result<Option<&'static str>> {
-	let tracker_policy = workflow.frontmatter().tracker();
-
-	if issue.has_label(tracker_policy.opt_out_label()) {
-		return Ok(Some("issue_opted_out"));
-	}
-	if issue.has_label(tracker_policy.needs_attention_label()) {
-		return Ok(Some("issue_needs_attention"));
-	}
-
-	Ok(None)
-}
